Revert "refactored out helper methods"
This reverts commit 6da2db1ca3bc1b1403c5009aa0765692ae226f4a.
